Imported upstream version '1.8.9001' of 'upstream'

+++ googletest/src/gtest-test-part.cc
@@ -70,7 +70,7 @@ const TestPartResult& TestPartResultArray::GetTestPartResult(int index) const {
     internal::posix::Abort();
   }
 
-  return array_[static_cast<size_t>(index)];
+  return array_[index];
 }
 
 // Returns the number of TestPartResult objects in the array.
